+++ app/Http/Controllers/UserController.php
@@ -5,14 +5,24 @@ namespace App\Http\Controllers;
 use App\Direccion;
 use App\FooterInfo;
 use App\HeaderFrontend;
+use App\SiteSocial;
 use App\TeamMember;
 use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Hash;
+use Session;
+use Image;
 
 class UserController extends Controller
 {
+    public function __construct(){
+        if(Session::has('auth_change_pass_user')){
+            Session::remove('auth_change_pass_user');
+        }
+    }
+
     public function index(){
         if(!Auth::check()){
             if(!session()->has('from')){
@@ -23,7 +33,8 @@ class UserController extends Controller
         $data = [
             'header' => HeaderFrontend::find(1),
             'footer' => FooterInfo::find(1),
-            'members' => TeamMember::all()
+            'members' => TeamMember::all(),
+            'siteSocials' => SiteSocial::all()
         ];
         return view('frontend.user', $data);
     }
@@ -33,20 +44,26 @@ class UserController extends Controller
         if($request->img){
             if($request->get('email') == $user->email){
                 $request->validate([
-                    'img' => 'required|image|mimes:jpeg,png,jpg|max:2048|dimensions:max_width=300,max_height=300',
-                    'name' => 'required|max:255'
+                    'img' => 'required|image|mimes:jpeg,png,jpg|max:2048|dimensions:min_width=300,min_height=300',
+                    'name' => 'required|max:50',
+                    'lastname' => 'required|max:50',
+                    'telefono' => 'required|max:50',
                 ]);
             }else{
                 $request->validate([
-                    'img' => 'required|image|mimes:jpeg,png,jpg|max:2048|dimensions:max_width=300,max_height=300',
-                    'name' => 'required|max:255',
+                    'img' => 'required|image|mimes:jpeg,png,jpg|max:2048|dimensions:min_width=300,min_height=300',
+                    'name' => 'required|max:50',
+                    'lastname' => 'required|max:50',
+                    'telefono' => 'required|max:50',
                     'email' => 'required|max:255|email|unique:users'
                 ]);
             }
-
+            //get image file
             $img = $request->file('img');
-
-            $imageName = time().'.'.$img->extension();
+            //get only filename
+            $img_filename = pathinfo($img->getClientOriginalName())['filename'];
+            //rename with date
+            $imageName = $img_filename.time().'.'.$img->extension();
 
             $imgResize = Image::make($img->path());
             $imgResize->fit(300,300, function($constraint) {
@@ -55,17 +72,23 @@ class UserController extends Controller
 
             $img_delete = 'images/uploads/users/'. $user->img;
             if(File::exists(public_path($img_delete))) {
-                File::delete($img_delete);
+                if($user->img != 'avatar.png'){
+                    File::delete($img_delete);
+                }
             }
             $user->img = $imageName;
         }else{
             if($request->get('email') == $user->email){
                 $request->validate([
-                    'name' => 'required|max:255',
+                    'name' => 'required|max:50',
+                    'lastname' => 'required|max:50',
+                    'telefono' => 'required|max:50',
                 ]);
             }else{
                 $request->validate([
-                    'name' => 'required|max:255',
+                    'name' => 'required|max:50',
+                    'lastname' => 'required|max:50',
+                    'telefono' => 'required|max:50',
                     'email' => 'required|max:255|email|unique:users'
                 ]);
             }
@@ -74,6 +97,8 @@ class UserController extends Controller
 
         $user->name = $request->get('name');
         $user->email = $request->get('email');
+        $user->lastname = $request->get('lastname');
+        $user->telefono = $request->get('telefono');
 
         $user->update();
 
@@ -98,6 +123,48 @@ class UserController extends Controller
 
         $user->direccions()->save($direccion);
 
+        return redirect('/cuenta');
+    }
+
+    public function changePassword(Request $request){
+        if(Session::has('auth_change_pass_user') && Session::get('auth_change_pass_user')){
+            $data = [
+                'header' => HeaderFrontend::findOrFail(1),
+                'footer' => FooterInfo::findOrFail(1),
+                'members' => TeamMember::all(),
+                'siteSocials' => SiteSocial::all()
+            ];
+            return view('frontend.templates.change-password', $data);
+        }else{
+            return redirect('/cuenta')->withErrors('Error al cambiar la contraseña');
+        }
+
+    }
+
+    public function checkPassword(Request $request){
+        if(Hash::check($request->get('oldpass'), Auth::user()->password)){
+            $request->session()->put('auth_change_pass_user', true);
+        }
+        return redirect(route('user.view.changepass'));
+    }
+
+    public function resetCheckPassword(Request $request){
+        if(Session::has('auth_change_pass_user')){
+            $request->session()->remove('auth_change_pass_user');
+        }
+    }
+
+    public function updatePassword(Request $request){
+        $request->validate([
+            'oldpass' => 'required',
+            'newpass' => 'required|min:8',
+            'repeatpass' => 'required|same:newpass'
+        ]);
+
+
+        $user = Auth::user();
+        $user->password = Hash::make($request->get('newpass'));
+        $user->save();
         return redirect('/cuenta');
     }
 }
